fix(npoint): correct pet time nang and power update
- Remove unnecessary cast in Mob.calSucManhTiemNang
- Fix pet time nang: remove incorrect doubling for charms and Buax2DeTu
- Use boolean flag for Buax2DeTu condition
- In addSMTN, add param <= 0 guard
- Remove direct master powerUp/tiemNangUp (recursive call handles)

+++ src/mob/Mob.java
@@ -236,7 +236,7 @@ public class Mob {
             tiemNang = 1;
         }
         if (pl.nPoint != null) {
-            tiemNang = (int) pl.nPoint.calSucManhTiemNang(tiemNang);
+            tiemNang = pl.nPoint.calSucManhTiemNang(tiemNang);
         } else {
             return 0;
         }
